feat: Generate region pages.
